Remove "dual buffer" logic in TCPConnection reading
I brought the basic idea for the buffer logic over from my work on Wallaroo and
that from TCPConnection. In the sense that Pony's TCPConnection has the
idea of a buffer that can grow over time and I carried that through all
my thinking. However, there's no good answer for "when do I make the
buffer bigger" if you are allocating a large buffer and chopping off of
it. The fundamental usage of memory is different than the growable,
truncate after read buffer.

This change drops us down to a single buffer size that is set and that
is what we will use. Couple issues that might need addressing:

* Yield is tied to buffer size

After reading in the buffer, we won't yield until all the data is
processed if we read the max amount. This may or may not be problematic.

* Buffer has to be at least "expect" size when doing framing

Trying to use expect to set to a value larger than the buffer size will
result in an error, so, you need to have a buffer that is as large as
your largest expected payload.

Neither is likely to be a real issue but I'm noting now.