Add epoch to open connection tokens (#135)
This should handle "many many connections have been opened" better
for the unlikely scenario that an id was to wrap around while some
"early token holders" are still open.

Now, we need to go through not just USize tokens but USize epochs
before we hit a conflict with overlapping ids.