Make connection limit more accurate (#138)
Keeping a count across _accept calls for opens we haven't seen yet
should be much more accurate. The tricky bit is if we get a close
for an open we never saw, we should drop the unseen opens number
but if we have seen the connection come back to us then we already
have dropped it.